Do not enable optional addons
The Nvidia and Mellanox addons are not necessary for a cluster; do not
enable them for now.

Also do not enable the etcd-defrag addon for now; it uses the same
version as the openstack-cluster chart by default which may be different
in our repo. This can be re-enabled when it is properly built in our CI.

Change-Id: I0b0941de5c2402f2416f6541d259ed0eb47d9168